requirements.txt
flask
youtube_transcript_api
numpy
pandas
transformers
nltk
scikit-learn
langdetect
mysql-connector-python
torch